Fix Cody Web demo after UI tabs update
