+++ vscode/webviews/tabs/index.ts
@@ -2,4 +2,5 @@ export { AccountTab } from './AccountTab'
 export { CommandsTab } from './CommandsTab'
 export { HistoryTab } from './HistoryTab'
 export { SettingsTab } from './SettingsTab'
-export { TabsBar, View } from './TabsBar'
+export { TabsBar } from './TabsBar'
+export { View } from './types'
