+++ vscode/webviews/tabs/types.ts
@@ -1,0 +1,8 @@
+export enum View {
+    Chat = 'chat',
+    Login = 'login',
+    History = 'history',
+    Account = 'account',
+    Commands = 'commands',
+    Settings = 'settings',
+}
